Added a USER-DEMSI dependence on KOKKOS.
Added a line in CMakeLists.txt to install all KOKKOS_HEADERS if
PKG_KOKKOS is on. This fixes the .h files not being installed.

Created a check on Kokkos to see if it is already initialized, and if
it is to use the same number of threads, numa regions, and ngpus, and
to only finalize Kokkos conditional on whether or not LAMMPS
initialized Kokkos.